% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eval_data.R
\name{generate_eval_data}
\alias{generate_eval_data}
\title{Generate evaluation data for a hub}
\usage{
generate_eval_data(hub_path, config_path, out_path, oracle_output)
}
\arguments{
\item{hub_path}{A path to the hub.}

\item{config_path}{A path to a yaml file that specifies the configuration
options for the evaluation.}

\item{out_path}{The directory to write the evaluation data to.}

\item{oracle_output}{A data frame of oracle output to use for the evaluation.}
}
\description{
Generate evaluation data for a hub
}
