% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Load and validate a predevals config file}
\usage{
read_config(hub_path, config_path)
}
\arguments{
\item{hub_path}{A path to the hub.}

\item{config_path}{A path to a yaml file that specifies the configuration
options for the evaluation.}
}
\value{
A list of configuration options for the evaluation.
}
\description{
Load and validate a predevals config file
}
